+++ plugins/config-edit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @powersync/cli-plugin-config-edit
 
+## 0.9.5
+
+### Patch Changes
+
+- @powersync/cli-core@0.9.5
+
+## 0.9.4
+
+### Patch Changes
+
+- @powersync/cli-core@0.9.4
+
 ## 0.9.3
 
 ### Patch Changes